feat: post 좋아요

+++ src/main/java/com/reborn/server/domain/community/dto/response/CommunityPostResponse.java
@@ -28,7 +28,7 @@ public class CommunityPostResponse {
     private String content;
     private String region;
     private String postImage;
-    private int likesCount;
+    private Long likesCount;
     private int commentsCount;
     private LocalDateTime createdAt;
     private List<String> interestTags;
